fix: Do not cache empty blocks in fetchAndCachePosts (#69)
# Purpose
The goal of this PR is to fix posts not showing correctly.

## Solution

`fetchAndCachePosts` was getting out of sync. 

In the case where `fetchAndCachePosts` is called for a block, but it
executes before the `content-watcher` sends the webhook notification, it
would cache an empty block.

Then the announcement comes from `content-watcher`, the cache skips the
block because it thinks it is already in the cache.

`fetchAndCachePosts` was modified so that empty blocks are not cached.

Developed with help from:
@JoeCap08055 @saraswatpuneet 

### Change summary

- `fetchAndCachePosts` was modified so that empty blocks are not cached.
- `npm audit fix` was used to update `package-lock.json` to fix
vulnerabilites.


## Steps to Verify
1. Startup SAT Backend and Frontend
2. Create a new post and wait about 30-60 seconds (for content-watcher
interval)
3. Refresh the frontend and verify that `My Posts` shows the post that
was just created.