feat: track graph operations by refId (#107)
# Purpose
The goal of this PR is to track pending graph operations by their
returned `referenceId`.

When submitting a follow/unfollow request, the payload now includes a
webhook URL that the `graph-service` will use to notify this app of the
completion status of the operation. The GraphWebhookService keeps track
of pending graph operations by their referenceId. A new endpoint, `GET
/graph/operation-status/{:refId}` is available for the frontend to query
the operation status.

Closes #51